requirements.txt
edge-tts
moviepy
Pillow
google-api-python-client
google-auth-oauthlib
requests
scipy
flask
gunicorn
psutil
